# docs/ODESwagger.yaml
swagger: '2.0'
info:
  title: ODE REST API
  description: >-
    Description of functions available from the Operational Data Environment
    (ODE) via RESTful interface. Ranges given follow interval notation, for
    example (0,1) or (0..1) means 0 &lt; x &lt; 1 and [0,1] or [0..1] means 0
    &le; x &le; 1.
  version: 0.0.1
host: yourhostname
schemes:
  - https
basePath: /
produces:
  - application/json
paths:
...
  /tim/query:
    post:
      summary: Query an RSU for set TIMs
      operationId: tim_query
      description: Query an RSU for set TIMs
      tags:
        - TIM
      parameters:
        - name: RSU
          in: body
          required: true
          schema:
            $ref: '#/definitions/RSU'
      responses:
        '200':
          description: List of messages in JSON form
          schema:
            $ref: '#/definitions/Status'
        '400':
          description: Error in body request
          schema:
            $ref: '#/definitions/Status'
        '408':
          description: Timeout
          schema:
            $ref: '#/definitions/Status'
        '500':
          description: Misc error
          schema:
            $ref: '#/definitions/Status'
...
definitions:
...
  RSU:
    type: object
    description: 'Optional field, needed for SNMP deposit to RSU.'
    properties:
      rsuIndex:
        type: integer
        description: Index of message storage on RSU.
      rsuTarget:
        type: string
        description: IPv4 address of RSU
      rsuUsername:
        type: string
        description: SNMP username (Not required if value is pre-configured in ODE properties)
      rsuPassword:
        type: string
        description: SNMP password (Not required if value is pre-configured in ODE properties)
      rsuRetries:
        type: integer
        description: Number of retries upon message send failure
      rsuTimeout:
        type: integer
        description: Timeout measured in milliseconds
...
  Status:
    type: object
    properties:
      success:
        type: boolean
      message:
        type: string